Swap FeedList and ItemListFragments
